requirements.txt
bcbio_gff==0.7.0
beautifulsoup4==4.11.1
biopython==1.83
dash==2.16.1
dash_bio==1.0.2
dash_bootstrap_components==1.5.0
dash_core_components==2.0.0
dash_html_components==2.0.0
dash_iconify==0.1.2
dash_mantine_components==0.14.4
Flask==2.2.5
Flask_Caching==2.3.0
Flask_Limiter==3.8.0
ip2geotools==0.1.6
Jinja2==3.0.3
logomaker==0.8
matplotlib==3.7.5
numpy==1.22.1
pandas==2.0.3
plotly==5.24.1
pygenomeviz==1.4.2
python-dotenv==1.0.1
Requests==2.32.3
setuptools==62.1.0
SQLAlchemy==1.4.36
urllib3==1.26.9
gunicorn>=20.1.0
pymysql==1.1.1